Configure a key pair for EC2 instance
This commit creates a new resource: 'aws_key_pair', which is an SSH pub key.
Then, it adds this key through the 'key_name' attribute to the EC2 resource
created previously. After that we can log into the EC2 by using the private
SSH key that corresponds to this public one.